+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 [unreleased]
+- Enhancement: Add right-click menu option to clear resume-at-line setting
 - Fixed: The 4th axis probing sequence for the z offset calibration (M469.5) was not passing pin diameter input through to the machine
 
 [2.1.0-RC1]